Fixed performance-check.jl so that it ignores jit compilation
performance-check.jl did not call the testing functions to "warm" the jit cache. If the tests are run from the command line "julia performance-check.jl" this meant that the timings would be including jit compilation. If the file was meant to be included from a juila prompt, this change is not necessary as you can include it multiple times, to remove the jit compillation from the timings.